requirements.txt
pandas
numpy 
flask
sqlalchemy
Flask-sqlalchemy
polygon-api-client
requests
flask-marshmallow
marshmallow-sqlalchemy
mysql-connector-python

